Removed xxx_uncompressed_size functions (since they were not constant-time and basically did decompression without writing).
Removed finish argument from xxx_inflate as it is not used in frameworks like this and was only being used for error detection (which was moved to xxx_inflate_end).

Changed finish argument of xxx_deflate to flush, which supports MSCOMP_NO_FLUSH (equivilent to old finish=false), MSCOMP_FINISH (equivilent to old finish=true), and MSCOMP_FLUSH which will try to get all current read input into the output and not buffer it (similar to finishing except the stream is not terminated).